Refactor purchase package following main
