+++ src/main/java/com/bh/rms/domain/aggregate/purchase/PurchaseService.java
@@ -1,10 +1,10 @@
-package com.bh.rms.domain.aggregate.purchase.service;
+package com.bh.rms.domain.aggregate.purchase;
 
 import com.bh.rms.domain.aggregate.material.exception.MaterialNotFoundException;
 import com.bh.rms.domain.aggregate.material.MaterialRepository;
 import com.bh.rms.domain.aggregate.purchase.Purchase;
 import com.bh.rms.domain.aggregate.purchase.PurchaseItem;
-import com.bh.rms.domain.aggregate.purchase.infra.PurchaseRepository;
+import com.bh.rms.domain.aggregate.purchase.PurchaseRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
