+++ src/main/java/com/bh/rms/domain/compositions/cost/CostService.java
@@ -2,7 +2,7 @@ package com.bh.rms.domain.compositions.cost;
 
 import com.bh.rms.domain.aggregate.material.Material;
 import com.bh.rms.domain.aggregate.purchase.PurchaseItem;
-import com.bh.rms.domain.aggregate.purchase.infra.PurchaseRepository;
+import com.bh.rms.domain.aggregate.purchase.PurchaseRepository;
 import com.bh.rms.domain.aggregate.material.MaterialRepository;
 import com.bh.rms.domain.aggregate.recipe.Recipe;
 import com.bh.rms.domain.aggregate.recipe.infra.RecipeRepository;
