+++ src/test/java/com/bh/rms/domain/aggregate/purchase/service/PurchaseServiceTest.java
@@ -4,8 +4,9 @@ import com.bh.rms.domain.aggregate.material.exception.MaterialNotFoundException;
 import com.bh.rms.domain.aggregate.material.MaterialRepository;
 import com.bh.rms.domain.aggregate.purchase.Purchase;
 import com.bh.rms.domain.aggregate.purchase.PurchaseItem;
+import com.bh.rms.domain.aggregate.purchase.PurchaseService;
 import com.bh.rms.domain.aggregate.purchase.exception.PurchaseNotFoundException;
-import com.bh.rms.domain.aggregate.purchase.infra.PurchaseRepository;
+import com.bh.rms.domain.aggregate.purchase.PurchaseRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
